Move `Structures.h` and `Robots.h` to "MapObjects" folder
There help with inclusion, and are not themselves header files for derived classes. As such, it makes sense to lift them up one folder level.
