+++ OPHD/MapObjects/Robots.h
@@ -1,0 +1,6 @@
+#pragma once
+
+// Incldues all Robots
+#include "Robots/Robodozer.h"
+#include "Robots/Robodigger.h"
+#include "Robots/Robominer.h"
