--- OPHD/MapObjects/Robots/Robots.h
@@ -1,6 +1,0 @@
-#pragma once
-
-// Incldues all Robots
-#include "Robodozer.h"
-#include "Robodigger.h"
-#include "Robominer.h"
