+++ OPHD/MapObjects/Structures.h
@@ -1,0 +1,41 @@
+#pragma once
+
+// Includes all Structures
+#include "Structures/Agridome.h"
+#include "Structures/AirShaft.h"
+#include "Structures/CargoLander.h"
+#include "Structures/CHAP.h"
+#include "Structures/ColonistLander.h"
+#include "Structures/CommandCenter.h"
+#include "Structures/Commercial.h"
+#include "Structures/CommTower.h"
+#include "Structures/FusionReactor.h"
+#include "Structures/HotLaboratory.h"
+#include "Structures/Laboratory.h"
+#include "Structures/MaintenanceFacility.h"
+#include "Structures/MedicalCenter.h"
+#include "Structures/MineFacility.h"
+#include "Structures/MineShaft.h"
+#include "Structures/Nursery.h"
+#include "Structures/Park.h"
+#include "Structures/SurfacePolice.h"
+#include "Structures/UndergroundPolice.h"
+#include "Structures/RecreationCenter.h"
+#include "Structures/Recycling.h"
+#include "Structures/RedLightDistrict.h"
+#include "Structures/Residence.h"
+#include "Structures/Road.h"
+#include "Structures/RobotCommand.h"
+#include "Structures/SeedFactory.h"
+#include "Structures/SeedLander.h"
+#include "Structures/SeedPower.h"
+#include "Structures/SeedSmelter.h"
+#include "Structures/Smelter.h"
+#include "Structures/SolarPanelArray.h"
+#include "Structures/SolarPlant.h"
+#include "Structures/StorageTanks.h"
+#include "Structures/SurfaceFactory.h"
+#include "Structures/Tube.h"
+#include "Structures/UndergroundFactory.h"
+#include "Structures/University.h"
+#include "Structures/Warehouse.h"
