--- OPHD/MapObjects/Structures/Structures.h
@@ -1,41 +1,0 @@
-#pragma once
-
-// Includes all Structures
-#include "Agridome.h"
-#include "AirShaft.h"
-#include "CargoLander.h"
-#include "CHAP.h"
-#include "ColonistLander.h"
-#include "CommandCenter.h"
-#include "Commercial.h"
-#include "CommTower.h"
-#include "FusionReactor.h"
-#include "HotLaboratory.h"
-#include "Laboratory.h"
-#include "MaintenanceFacility.h"
-#include "MedicalCenter.h"
-#include "MineFacility.h"
-#include "MineShaft.h"
-#include "Nursery.h"
-#include "Park.h"
-#include "SurfacePolice.h"
-#include "UndergroundPolice.h"
-#include "RecreationCenter.h"
-#include "Recycling.h"
-#include "RedLightDistrict.h"
-#include "Residence.h"
-#include "Road.h"
-#include "RobotCommand.h"
-#include "SeedFactory.h"
-#include "SeedLander.h"
-#include "SeedPower.h"
-#include "SeedSmelter.h"
-#include "Smelter.h"
-#include "SolarPanelArray.h"
-#include "SolarPlant.h"
-#include "StorageTanks.h"
-#include "SurfaceFactory.h"
-#include "Tube.h"
-#include "UndergroundFactory.h"
-#include "University.h"
-#include "Warehouse.h"
