+++ OPHD/StructureCatalogue.cpp
@@ -1,7 +1,7 @@
 #include "StructureCatalogue.h"
 
 #include "StorableResources.h"
-#include "MapObjects/Structures/Structures.h"
+#include "MapObjects/Structures.h"
 
 #include <string>
 #include <stdexcept>
